add script to enable continuous atomic host delivery from rawhide
allow making atomic host from the rawhide buildroot on aarch64, armhfp
and x86_64 from the unsigned content in the rawhide buildroot. This
is to be enabled in a cron job to run every 15 minutes so that as soon
as anything going into atomic host changes there is the ability for
people to test and fix anything before it hits rawhide.  We will detect
sooner when things break. This content is intended purely for testing
purposes.

Signed-off-by: Dennis Gilmore <dennis@ausil.us>