refactor block_retired to handle pkgdb namespaces
Pkgdb has added the concept of "namespaces" allowing non-rpm
artifacts to live in pkgdb and DistGit. The namespace "rpms" is
default in pkgdb for backwards compatibility. This patch enables the
handling of non-rpm namespaces and adds a mapping specifically for
the "docker" namespace in order to handle the blocking of Fedora
Container Layered Images. The default namespace will remain "rpms"
for backwards compatibility (as it is with pkgdb also) and other
mappings can easily be added in the future.

Signed-off-by: Adam Miller <maxamillion@fedoraproject.org>