block_retired.py: Do not login to the koji session
The koji calls to list the packages in a tag are unauthenticated
so there is no need to login. The login call has been failing due
the session options not being passed to the session. However to
complete the intended function we do not need to login, so lets
not.

Signed-off-by: Dennis Gilmore <dennis@ausil.us>