Add container namespace mapping
The container namespace in DistGit/PagkageDB has migrated from
"docker" to "container" in order to keep it more generic than
a specific implementation. This update adds support for the new
namespace mappings. Note that we will no longer add mappings to the
"docker" namespace mappings and will eventually retire it from this
script once they are no longer needed.

Signed-off-by: Adam Miller <maxamillion@fedoraproject.org>