Remove this weird loop.
It looks like it is left over as a mistake from
e8c509074e8dd0dc932565884113505f61ebac33.

Previously it seemed to loop over the primary and secondary koji
instance, but in e8c509074e it was mistakenly changed to loop over the
primary architectures themselves -- which doesn't work.

Since everything is rolled into primary koji now, this is all we should need.

Signed-off-by: Ralph Bean <rbean@redhat.com>