Don't sort commiters
Sorting users by alphabet gives higher precedence to users by their nick
name. This does not seems to be right, since user with nickname starting
by "a" is probably not any better then user with nick name starting by
"b".

OTOH, it seems that Pagure returns committers in order as they gained
the commit privileges (not 100% sure), which to me is much better
metrics, since if nothing else, the earlier the committer might have
more experiences with the package.

Signed-off-by: Vít Ondruch <vondruch@redhat.com>