signoff on commits for mass rebuild
with pagure on dist-git pkg owners can enable a toggle switch to
require all commits be signed off on. as per Developer Certificate
of Origin (https://developercertificate.org/) it is good behaviour
anyway. With a question being raised already in a pagure issue over
using the feature to block mass rebuild builds
https://pagure.io/pagure/issue/2459

Signed-off-by: Dennis Gilmore <dennis@ausil.us>