Update sop_two_week_atomic to include latest AutoCloud info
In the past, AutoCloud was slated to be the authoritative place for
testing. However, this never actually panned out and there are now
some work arounds required at times. This pull request updates the
SOP to include documentation of those work arounds as well as
a description of current process and when work arounds are needed.

Signed-off-by: Adam Miller <maxamillion@fedoraproject.org>