atomic: operate directly on prod ostree repo
Patrick informed me that we can (and should) now
be operating directly on the prod ostree repo instead
of the compose/updates repo. This means we need to add
back in passing of release arg (to fill in to ATOMIC_HOST_DIR
variable) and also go back to updating the summary file.

Signed-off-by: Dusty Mabe <dusty@dustymabe.com>