Ensure the namespace is listed in the file created during mass-branching
This file is generated when we branch the components in PDC for the new
branch.
Currently it was only returning the name of the artifact without their
namespace, so if you have 'cockpit' you had no idea if you're speaking
about the rpm, the container or the module.
This commit fixes that.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>