Make find_failures.py show all maintainer
Previously the find_failures.py would use only the first result
provided by pagure, which caused some package maintainers to fail to
notice problems from the mass rebuild.

In addition, the script would make one API call per failed package,
which was quite slow.

This patch converts the script to show list packages under each
maintainer, and to use the data which is cached (and updated hourly) for
just this use case.

Instead of this, used the data which is cached for the express this, and
list packages under each maintainer.

For me, the runtime drops from 2 minutes 41 seconds to 18 seconds.  The
resulting list is of course a bit longer and many packages are listed
under multiple usernames but the end result looks much the same.

Signed-off-by: Jason Tibbitts <tibbs@math.uh.edu>