block_retired: Block packages which are not active in PDC
This is opposite to what was happening, we were blocking packages which
are not active... But some packages might not even be recorded in PDC
for specific branch.

Let's just do opposite, find all unblocked packages in Koji, all active
packages in PDC and block what is not in the list.

Signed-off-by: Igor Gnatenko <i.gnatenko.brain@gmail.com>