Mass rebuild: Make sure we're on the master branch no matter the default
Pagure dist-git repos can have different default branches than master,
but during the mass rebuild, we always want to use the master branch.

This clones the repository on master even if it is not the default,
and it also fails (=skips the package) if there is no master branch.

This prevents cases like https://pagure.io/releng/issue/9213

Signed-off-by: everybody <all@the.people>