Look at local and remote branches separately
When running in a user clone of a dist-git repo, we want to check if
the *remote* branch can be deleted:
$ fedpkg clone foo
$ cd foo
$ .../distgit-commit-unused.py origin/f32

In that case we only check if origin/f32 is contained any remote branch.

When run by releng, I expect that this will be run locally, so
$ cd .../foo
$ .../distgit-commit-unused.py f32

In that case, there should be no remote branches, but even if there
were, we do not care about them.