Add a script that returns who maintains what and allows to retire them
Basically, this script lists who maintains what (and thus replaces the
"who_maintains_what.py" script that will be removed from this repo) and
if given the ``--retire`` argument, it will orphan all the packages for
which the packager(s) is/are main admin, it will remove the ACL for the
packager(s) on packages they have commit on and will reset their watch
status for everything they watch.

This script can be used when someone goes MIA and we want to mass-orphan
their packages. It can also be used when an account no longer has a valid
bugzilla account associated with their email in FAS.

Finally, the script can also only lists/acts on packages the packagers
maintain (via the ``--maintain`` argument) or only packages they watch
(via the ``--watch``argument). Allowing to better tune the action
performed.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>