When querying the list of packages maintained by people, also query dist-git itself
So far we only relied on the pagure_bz.json file to tell us what
someone was involved with. However, if someone set their watch
status to "unwatch" on a package, that package will not show up
in pagure_bz.json.
So we'll now be querying both that file and dist-git itelf directly.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>