retire_packagers: Use the dist-git's specific endpoint to orphan packages
This endpoint is specific to dist-git and does more than the usual
endpoint by ensuring that the former main admin is removed from the
project as well as resetting their watch status (so we no longer have
to make an extra request for this).

It also announces that the package was orphaned on fedora-messaging and
we are working on automatically notifying the devel list when a package
is orphaned so people are aware of it as soon as possible and can decide
if they want to pick it up or not.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>