Fix recursion query in find_unblocked_orphans.py
There is a check in find_dependent_packages() to prevent packages reported as
depended upon when the dependency can be resolved by another package.

For some reason, for (some?) Python packages, this check triggered even for
packages from the same component. This usually stopped the dependency query on
the first level even if many other packages depended recursively.

This fixes the problem by explicitly not triggering the check for packages
from the same component.

Signed-off-by: Miro Hrončok <miro@hroncok.cz>