Remove stray dash in commit messages for mass rebuild
The dash is appropriate in %changelog, because it's part of the
formatting there. But there is no reason to include it in git commit
messages.

Now that %changelog is generated from git log by rpmautospec, this is
causing problems because we try to remove the extra dash and it's not
always clear if it should be removed or not, see the discussion in
https://pagure.io/fedora-infra/rpmautospec/pull-request/242 .

It turns out rpmdev-bumpspec will insert a dash on its own, so we can
use the same string in both places and get the expected formatting.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>