FailsToInstall bugzillas: Clear all needinfos on closed bugzillas
Previously, the bugzillas were closed, but the needinfos remained hanging.
That means maintainers were rightfully irritated
that they need to remove the needinfos themselves afterward.

In an ideal world, the maintainers would close the bugzillas themselves,
clearing the needinfos. But with some large fallout from mass rebuilds etc.
it is probably too much to expect.

This way, if they just wait it out, everything will be good once the issue is solved.

The clearing of needinfos takes time and makes the script run longer
when there are many bugzillas to be closed,
but machine-time is cheap compared to people-time.

Signed-off-by: Miro Hrončok <miro@hroncok.cz>