Add mail templates for mail announcements releng sends
These are emails releng sends every cycle.
In the past we simply copied them from previous announcements.
However, that makes it harder to fix them for suggestions or allow
others to help us make them better via PRs.

So, check them in here and use them from here from now on.

For now I just replaced versions with N and dates with YYYY-MM-DD type
of things. We may want to change that to something thats easier to
template at some point, but this at least is a good first cut.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>