fbtfs-fti template: Fix mock command to reproduce BR
This makes sure that the mock chroot's `chroot_setup_command` is set to
install `@build` instead of `@buildsys-build`, which is not available in
the BR repo, in addition to disabling all repos except `local`. This
fixes an issue where the mock chroot is unable to bootstrap itself due
to @buildsys-build not existing in the koji buildroot repo.

Signed-off-by: Maxwell G <gotmax@e.email>