end-of-life template: Explicitly mention no more security announcements
The current template mentions that no more updates will be provided
for EOL Fedora releases.

This could create the (wrong) impression that the code is still evaluated
for security risks and that just the updates won't happen anymore.

To avoid any misunderstandings:
Add an explicit statement that also no more security announcements will
be made.

Patch update 2023-03-04:
Spelling update: Use fN for the Fedora release number.

Example:
https://bugzilla.redhat.com/show_bug.cgi?id=2030945
Only Fedora 34 and Fedora 35 are discussed. This doesn't mean that
Fedora 33 is not affected: Fedora 33 was already EOL, thus Fedora 33
was not evaluated.

Signed-off-by: Manfred Spraul <manfred@colorfullife.com>