critpath.py: let Python do the unicode work for us
This stuff with writing the file in binary mode and using
`.encode()` on the strings we're going to put into it smells
like an attempt to make it work the same in Python 2 and 3, and
indeed it comes from the Python 3 support commit. We don't care
about Python 2 any more, so let's just let Python do the work
here, by opening the file in text mode with utf-8 encoding
specified, and dropping the use of `.encode()`.

I think this is all fairly academic as we don't allow non-ASCII
characters in package names anyhow.

Signed-off-by: Adam Williamson <awilliam@redhat.com>