critpath.py: Explicitly pass variables to expand_dnf_critpath
`expand_dnf_critpath` uses the variables `url` and `arch` but
they are not passed to it. This worked because the variables are
actually declared in global scope and thus available to the
function, but relying on this is a bad idea and it'd break if we
refactor the code to not do a lot of work in global scope.

Signed-off-by: Adam Williamson <awilliam@redhat.com>