critpath.py: revise and fix URL arguments
This completely revises the logic for determining repo URLs and
handling alternative arguments. The existing code had outdated
hacks for running on nightly composes which haven't worked since
at least 2017. Throw those out, and refactor so you can pass
alternate URLs to the top-level fedora/linux and fedora-secondary
directories for working with repos (e.g. for using a different
mirror or a local filesystem if you have the repos there), *or*
you can pass a URL to the /compose directory of a non-split
Pungi compose, which overrides the other two settings if set.

Signed-off-by: Adam Williamson <awilliam@redhat.com>