critpath.py: restore update repository consideration for stables
The port to DNF dropped the inclusion of update repositories
when calculating critical path for stable releases. This meant
we couldn't change the critical path for a stable release, as
any changes go to the comps for the updates repo, not the frozen
release repo. With this change we should be able to update the
critpath for stable releases again. It also means the critpath
calculation will change to reflect changes of the dependencies
of critpath packages in updates.

Signed-off-by: Adam Williamson <awilliam@redhat.com>