spam-o-matic: set dnf `Base.conf.arch` to fix reports (#7931)
Thanks to mhroncok for suggesting this. @kevin pointed out that
reports were still completely wrong on some arches. It seems
that dnf filtering by provides needs this `Base.conf.arch`
config option set to the relevant arch to work properly. With
this set, generating a report for a different arch from the
'native' arch of the system generating the report should work
properly.

Signed-off-by: Adam Williamson <awilliam@redhat.com>