critpath.py: get release data from Bodhi, add an 'all' mode
This enhances critpath.py to get release data (what releases are
currently stable, whether there's a Branched and what its release
number is if so) from Bodhi instead of having it hardcoded.
Hardcoding is simpler but absolutely inevitably goes stale. In
the overall critpath plan I'm working towards this script will
be called regularly by Bodhi anyway, so it makes sense for it to
get its release data from Bodhi.

This also adds an 'all' mode which generates critpath data for
all current/pending releases (as determined from the Bodhi data)
and outputs it under the filenames expected by the new support
for reading such files in Bodhi 7.x.

Signed-off-by: Adam Williamson <awilliam@redhat.com>