mass-tag.py: Use rpm vercmp for determining what to tag over
This should fix the issue we had with tagging back builds in the latest
mass rebuild where the script tagged over a number of packages that had
actually gotten rebuilt in rawhide in the mean time.

Previously, the logic was to compare task start times, but the way it
found the "newer" task failed to take into account that not all builds
are done using XX-candidate/rawhide targets and failed to find builds
done in self-service side tags.

This commit simplifies the logic in the script and makes it just use rpm
vercmp to compare the latest builds between the source and target tags
to determine what to tag over, hopefully making it all less error prone.

Signed-off-by: Kalev Lember <klember@redhat.com>