critpath.py: add some additional critical path groups
These have been added to comps over time but weren't "active"
because they hadn't been added to this script. Also added a note
that the critpath groups *must* be kept synced with the decision
contexts in the greenwave policy (Bodhi will blow up if an update
is in a critpath group that there is no greenwave policy for).

Signed-off-by: Adam Williamson <awilliam@redhat.com>