mail-templates: adjust text about git
'git pull --rebase' is a rather circuitous (and potentially dangerous)
way to pick up a new remote branch. Since the branch was newly created
upstream, we know that it's not tracked locally yet, so we can just
use 'git fetch'. Add '-v' to make it more explicit to the user that a
new upstream branch is available.

Also use https:// for the link and move it to the links section
following the usual style.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>