critpath.py: handle the 'release gap' problem
There's this infamous issue with using release data from Bodhi
(as this script does) - a new release is marked 'stable' in
Bodhi several days before it's actually accessible in the normal
location for 'stable' releases in the mirror system (even on
dl.fp.o). If you just trust Bodhi to tell you when a release is
stable and assume you can find it in the stable path, for a few
days each cycle this will fail. I somehow failed to notice this
would be a problem for critpath.py when adjusting it to use Bodhi
data and be run daily, but...obviously it is:
https://pagure.io/fedora-infrastructure/issue/11247
This is a somewhat hacky way to handle it, but it's quite easy
and doesn't require too much change to the code.

Signed-off-by: Adam Williamson <awilliam@redhat.com>