Remove stray Signed-off-by in mass rebuild commits
All Fedora contributors must sign FAS to create an account and make
contributions, so an additional Signed-off-by is not needed. It is
not used for commits in dist-git.

In fact, having a Signed-off-by is problematic: Signed-off-by by
itself does not define a meaning, the meaning must be documented
externally. But Fedora does not provide such a document, so the
signature is meaningless. And if we were to define such a document in
the future, it would not cover previous signatures, because we'd have
no way to know if that's what the signer meant. This would leave to
massive confusion.

Also, those releng commits are highly visible to all contributors, and
make it likely the new contributors will copy this.

Drop the siganture to avoid noise and confusion.

Signed-off-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>